first version of admin dashboard, authentication missing
