+++ package.json
@@ -4,6 +4,8 @@
   "description": "Backend GraphQL server for Giveth originally forked from Topia",
   "main": "./dist/index.js",
   "dependencies": {
+    "@admin-bro/express": "^3.1.0",
+    "@admin-bro/typeorm": "^1.4.0",
     "@sentry/node": "^6.2.0",
     "@sentry/tracing": "^6.2.0",
     "@types/bcryptjs": "^2.4.2",
@@ -18,6 +20,7 @@
     "@types/node": "^14.14.31",
     "@types/uuid": "^7.0.4",
     "@uniswap/sdk": "^3.0.3",
+    "admin-bro": "^4.0.1",
     "analytics-node": "^4.0.1",
     "apollo-server": "^2.21.0",
     "apollo-server-express": "^2.21.0",
@@ -33,6 +36,7 @@
     "eth-sig-util": "^3.0.1",
     "ethers": "^5.0.31",
     "express": "^4.17.1",
+    "express-formidable": "^1.2.0",
     "express-session": "^1.17.1",
     "form-data": "^3.0.1",
     "graphql": "^15.5.0",
